--- src/components/PropsAndPropTypes.jsx
@@ -1,20 +1,0 @@
-import Props from './Props';
-
-export default function PropsAndPropTypes() {
-  return (
-    <>
-      <header>
-        <h1>Props and PropTypes</h1>
-        <p>
-          Props and PropTypes are two core concepts in React. They are used to
-          define the shape of a React component. They play a key role in any
-          React application as they help to render the component based on the
-          given data.
-        </p>
-      </header>
-      <main>
-        <Props />
-      </main>
-    </>
-  );
-}
